Sprint 39

+++ src/services/src/Demoulas.ProfitSharing.Common/Contracts/Request/Distributions/CreateDistributionRequest.cs
@@ -1,7 +1,7 @@
 ﻿namespace Demoulas.ProfitSharing.Common.Contracts.Request.Distributions;
 public record CreateDistributionRequest
 {
-    public int BadgeNumber { get; set; }
+    public long BadgeNumber { get; set; }
     public char StatusId { get; set; }
     public char FrequencyId { get; set; }
     public int? PayeeId { get; set; }
